Move printing to common module
